refactor: Extract common Firestore model classes
FirestoreTrainedExercise is Firestore representation of the model class
"TrainedExercise" and allows for an easier time parsing Firestore fields

Since some repository classes both depend on this model class, it's now
extracted to its own file instead of being private in the
dependant classes
