+++ app/src/main/java/com/haidoan/android/stren/core/datasource/RoutinesFirestoreDataSource.kt
@@ -3,10 +3,8 @@ package com.haidoan.android.stren.core.datasource
 import com.google.firebase.firestore.DocumentId
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.QuerySnapshot
-import com.haidoan.android.stren.core.model.Exercise
+import com.haidoan.android.stren.core.datasource.model.FirestoreTrainedExercise
 import com.haidoan.android.stren.core.model.Routine
-import com.haidoan.android.stren.core.model.TrainedExercise
-import com.haidoan.android.stren.core.model.TrainingMeasurementMetrics
 import kotlinx.coroutines.tasks.await
 import timber.log.Timber
 import javax.inject.Inject
@@ -53,54 +51,6 @@ class RoutinesFirestoreDataSource @Inject constructor() : RoutinesRemoteDataSour
                 trainedExercises = trainedExercises
             )
         }
-    }
-
-    private data class FirestoreTrainedExercise(
-        val exerciseCategory: String = "Undefined",
-        val exerciseId: String = "Undefined",
-        val name: String = "Undefined",
-        val note: String = "Undefined",
-        val trainingSets: Map<String, Long> = mapOf(),
-    ) {
-        fun asTrainedExercise(): TrainedExercise {
-            val firestoreTrainingSets = this.trainingSets
-            val trainingSets: MutableList<TrainingMeasurementMetrics> = mutableListOf()
-
-            when (this.exerciseCategory) {
-                ExerciseCategoryWithSpecialMetrics.CARDIO.fieldValue -> {
-                    firestoreTrainingSets.forEach {
-                        trainingSets.add(
-                            TrainingMeasurementMetrics.DistanceAndDuration(
-                                it.key.toLong(),
-                                it.value.toDouble()
-                            )
-                        )
-                    }
-                }
-                ExerciseCategoryWithSpecialMetrics.STRETCHING.fieldValue -> {
-                    firestoreTrainingSets.forEach {
-                        trainingSets.add(TrainingMeasurementMetrics.DurationOnly(it.value))
-                    }
-                }
-                else -> {
-                    firestoreTrainingSets.forEach {
-                        trainingSets.add(TrainingMeasurementMetrics.WeightAndRep(it.key, it.value))
-                    }
-                }
-            }
-            Timber.d("trainingSets: $trainingSets")
-            return TrainedExercise(
-                exercise = Exercise(
-                    id = this.exerciseId,
-                    name = this.name,
-                    belongedCategory = this.exerciseCategory
-                ), trainingSets = trainingSets.toList()
-            )
-        }
-    }
-
-    private enum class ExerciseCategoryWithSpecialMetrics(val fieldValue: String) {
-        CARDIO("Cardio"), STRETCHING("Stretching"),
     }
 }
